data: Move selected change loading
Hunk navigation owns fetching, filtering, and recalculating the live
selected hunk for command workflows.

Pathless command actions still read selected-change objects through
hunk_tracking, tying command validation to navigation even when no hunk
advance happens.

This commit begins separating selected-change access by moving loading and
stale-hunk validation into data.selected_change.loading and updating
command importers.

Subsequent commits will move the selected-change loading tests and guard
hunk_tracking from exposing those readers.
