+++ src/git_stage_batch/commands/include.py
@@ -57,6 +57,8 @@ from ..data.text_lifecycle_detection import detect_empty_text_lifecycle_change
 from ..data.hunk_tracking import (
     apply_line_level_batch_filter_to_cached_hunk,
     fetch_next_change,
+)
+from ..data.selected_change.loading import (
     load_selected_change,
     require_selected_hunk,
 )
