+++ src/git_stage_batch/commands/suggest_fixup.py
@@ -8,7 +8,7 @@ import sys
 
 from ..core.line_selection import parse_line_selection
 from ..data.file_review.fingerprints import compute_current_file_review_diff_fingerprint
-from ..data.hunk_tracking import require_selected_hunk
+from ..data.selected_change.loading import require_selected_hunk
 from ..data.file_hunk_display import render_file_as_single_hunk
 from ..data.line_state import load_line_changes_from_state
 from ..data.selected_change.store import get_selected_change_file_path
